requirements.txt
pbr
pyusb>=1.1.1
paho.mqtt.client
